backend/requirements.txt
# Dépendances principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de données
sqlalchemy==2.0.23
alembic==1.13.1

# Apprentissage par renforcement
stable-baselines3==2.0.0
gymnasium==0.29.1
pettingzoo==1.24.3

# Environnements Pac-Man existants (intégration)
# Les environnements sont dans src/pacman_env/

# Traitement d'images pour la visualisation
pillow==10.1.0

# Utilitaires
numpy==1.26.2
pandas==2.1.4
matplotlib==3.8.2

# Développement et tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Export de modèles
onnx==1.15.0
onnxruntime==1.16.3
torch==2.1.2

# Monitoring
psutil==5.9.7

# Sérialisation
pyyaml==6.0.1

# Note: Python 3.14 est requis (spécifié dans pyproject.toml du projet principal)